Move messageSubmissionSizeChange() to OCTokenSpec
